Changes to game logic
Changing format of gameBoard.su so it's more visibly pleasing in text
format. Easier to identify errors.
isMoveValid method is more lax. Will only return false if the tile is
uneditable.
Adding boolean isEditable to tile for pieces that are from the given
board vs what a user puts. This is so if a user makes a mistake then
they can overwrite their mistake.  Getting rid of pattern regex.